project: move telemetry
