--- source/telemetry/telemetry_logging/__init__.py
@@ -1,9 +1,0 @@
-"""
-To configure the logging module use the submodule `logging_configuration.py`.
-
-The following are the generally available hooks for logging from code.
-"""
-
-from .logger import Logger
-from .decorators import use_span
-from .span_recording import span_record_exception
